requirements.txt
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dimod==0.9.2
dwave-cloud-client==0.7.2
dwave-drivers==0.4.4
dwave-hybrid==0.4.3
dwave-inspector==0.2.1
dwave-inspectorapp==0.2.1
dwave-neal==0.5.4
dwave-networkx==0.8.6
dwave-ocean-sdk==2.2.0
dwave-qbsolv==0.3.0
dwave-system==0.9.5
dwave-tabu==0.3.0
dwavebinarycsp==0.1.2
entrypoints==0.3
Flask==1.1.2
homebase==1.0.1
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
json5==0.9.4
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.1.4
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.1.3
minorminer==0.1.9
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.1
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
notebook==6.0.3
numpy==1.18.1
ortools==7.6.7691
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
penaltymodel==0.16.3
penaltymodel-cache==0.4.1
penaltymodel-lp==0.1.3
penaltymodel-mip==0.2.3
pexpect==4.8.0
pickleshare==0.7.5
plucky==0.4.3
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.12.2
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyqubo==0.4.0
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pyzmq==18.1.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.23.0
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.1.0
